CHAD-17088: zwave-garage-door-opener lazy load subdrivers
