+++ drivers/SmartThings/zwave-garage-door-opener/src/ecolink-zw-gdo/fingerprints.lua
@@ -1,0 +1,9 @@
+-- Copyright 2025 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
+-- Ecolink garage door operator
+local ECOLINK_GARAGE_DOOR_FINGERPRINTS = {
+    {manufacturerId = 0x014A, productType = 0x0007, productId = 0x4731},
+}
+
+return ECOLINK_GARAGE_DOOR_FINGERPRINTS
