+++ drivers/SmartThings/zwave-garage-door-opener/src/mimolite-garage-door/fingerprints.lua
@@ -1,0 +1,8 @@
+-- Copyright 2025 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
+local MIMOLITE_GARAGE_DOOR_FINGERPRINTS = {
+  { manufacturerId = 0x0084, productType = 0x0453, productId = 0x0111 } -- mimolite garage door
+}
+
+return MIMOLITE_GARAGE_DOOR_FINGERPRINTS
